+++ src/tests/test_filesystem.py
@@ -105,3 +105,151 @@ def test_idbfs_persist_code(selenium_standalone):
     )
     # remove file
     selenium.run_js(f"""pyodide.FS.unlink("{mount_dir}/test_idbfs/__init__.py")""")
+
+
+@pytest.mark.xfail_browsers(
+    node="Not available", firefox="Not available", safari="Not available"
+)
+def test_nativefs_dir(request, selenium_standalone):
+
+    # Note: Using *real* native file system requires
+    # user interaction so it is not available in headless mode.
+    # So in this test we use OPFS (Origin Private File System)
+    # which is part of File System Access API but uses indexDB as a backend.
+
+    if request.config.option.runner == "playwright":
+        pytest.xfail("Playwright doesn't support file system access APIs")
+
+    selenium = selenium_standalone
+
+    selenium.run_js(
+        """
+        root = await navigator.storage.getDirectory();
+        dirHandleMount = await root.getDirectoryHandle('testdir', { create: true });
+        testFileHandle = await dirHandleMount.getFileHandle('test_read', { create: true });
+        writable = await testFileHandle.createWritable();
+        await writable.write("hello_read");
+        await writable.close();
+        fs = await pyodide.mountNativeFS("/mnt/nativefs", dirHandleMount);
+        """
+    )
+
+    # Read
+
+    selenium.run(
+        """
+        import os
+        import pathlib
+        assert len(os.listdir("/mnt/nativefs")) == 1, str(os.listdir("/mnt/nativefs"))
+        assert os.listdir("/mnt/nativefs") == ["test_read"], str(os.listdir("/mnt/nativefs"))
+
+        pathlib.Path("/mnt/nativefs/test_read").read_text() == "hello_read"
+        """
+    )
+
+    # Write / Delete / Rename
+
+    selenium.run(
+        """
+        import os
+        import pathlib
+        pathlib.Path("/mnt/nativefs/test_write").write_text("hello_write")
+        pathlib.Path("/mnt/nativefs/test_write").read_text() == "hello_write"
+        pathlib.Path("/mnt/nativefs/test_delete").write_text("This file will be deleted")
+        pathlib.Path("/mnt/nativefs/test_rename").write_text("This file will be renamed")
+        """
+    )
+
+    entries = selenium.run_js(
+        """
+        await fs.syncfs();
+        entries = {};
+        for await (const [key, value] of dirHandleMount.entries()) {
+            entries[key] = value;
+        }
+        return entries;
+        """
+    )
+
+    assert "test_read" in entries
+    assert "test_write" in entries
+    assert "test_delete" in entries
+    assert "test_rename" in entries
+
+    selenium.run(
+        """
+        import os
+        os.remove("/mnt/nativefs/test_delete")
+        os.rename("/mnt/nativefs/test_rename", "/mnt/nativefs/test_rename_renamed")
+        """
+    )
+
+    entries = selenium.run_js(
+        """
+        await fs.syncfs();
+        entries = {};
+        for await (const [key, value] of dirHandleMount.entries()) {
+            entries[key] = value;
+        }
+        return entries;
+        """
+    )
+
+    assert "test_delete" not in entries
+    assert "test_rename" not in entries
+    assert "test_rename_renamed" in entries
+
+    # unmount
+
+    files = selenium.run(
+        """
+        import os
+        os.listdir("/mnt/nativefs")
+        """
+    )
+
+    assert "test_read" in entries
+    assert "test_write" in entries
+    assert "test_rename_renamed" in entries
+
+    selenium.run_js(
+        """
+        await fs.syncfs();
+        pyodide.FS.unmount("/mnt/nativefs");
+        """
+    )
+
+    files = selenium.run(
+        """
+        import os
+        os.listdir("/mnt/nativefs")
+        """
+    )
+
+    assert not len(files)
+
+    # Mount again
+
+    selenium.run_js(
+        """
+        fs2 = await pyodide.mountNativeFS("/mnt/nativefs", dirHandleMount);
+        """
+    )
+
+    # Read again
+
+    selenium.run(
+        """
+        import os
+        import pathlib
+        assert len(os.listdir("/mnt/nativefs")) == 3, str(os.listdir("/mnt/nativefs"))
+        pathlib.Path("/mnt/nativefs/test_read").read_text() == "hello_read"
+        """
+    )
+
+    selenium.run_js(
+        """
+        await fs2.syncfs();
+        pyodide.FS.unmount("/mnt/nativefs");
+        """
+    )